include "libLoam/c/ob-sys.h" instead of <sys/time.h>
Addresses some issues related to "struct timeval" that were
raised by pull request #2.

Despite the original comment in the code, "struct timeval" is not
supposed to be defined in <stdlib.h>.

https://chatgpt.com/share/687182ec-5150-8006-866d-5dd2908008b7

On POSIX, "struct timeval" is supposed to be defined in
<sys/time.h>.  However, <sys/time.h> does not exist on Windows.
Therefore, the patch originally provided in #2 will not work on
Windows.

Instead, Windows defines "struct timeval" in <WinSock2.h>.

Therefore, the correct solution is to include
"libLoam/c/ob-sys.h".  ob-sys.h includes <sys/time.h> on POSIX,
and <WinSock2.h> on Windows.  ob-sys.h was created specifically
to abstract these sorts of system differences.