Hardcode version number to 5.6.0
Previously, it was trying to get the version number automatically from
git tags, which wasn't working, because we didn't have any tags in
this repo yet.

We could have fixed this by adding tags to the repo, but I'm not sure
that's what we want, since that would mean that the code wouldn't know
its version number if it is not being built from git.  (For example,
if we were to distribute source tarballs.)