fix cmake std::shuffle detection when compiler defaults to < C++-11.
CMakeLists.txt uses cmake's "check_cxx_source_compiles()" facility to
determine whether e.g. std::shuffle() is available; the test tries to
compile a program that references the function in question, and if it
succeeds then that function is known to be available. The attempted
compilation uses whatever version of C++ is the underlying compiler's
default, however -- irrespective of the C++ version that'll be used
for the actual compile that cmake is setting up for (!) -- and on OSX
12's clang-version-13 the default C++ is not even C++-11 (you'll find
the __cplusplus macro set to 199711L) and so the test code (as added
in commit 418d4147), which includes the line
      std::vector<int> v = {1,2,3,4,5};
, does not compile... but because of the initializer-list style that
only appears with C++-11 (whereas the line that exercises std::shuffle()
does not result in an error). Nonetheless, since the compiler test
fails, std::shuffle is marked as unavailable.

This patch primitivizes the C++ use (simply by declaring the variable
v first and then filling it via repeated application of push_back()),
and so now the test accurately detects the presence or absence of
std:shuffle().

Really, we could just excise the initializer-list part of the code
and live out our test-compiley life with an empty vector, but it is
not the place of this patch to alter the sense or intent of the code
it patches.