Update CHANGELOG.md for #17
Also, bump version number from 5.6.0 to 5.6.1.

I'm still figuring this out, but I've been thinking about using a
version numbering scheme where releases are even numbers, and
development versions are odd numbers.  This makes it possible to tell
if someone just grabbed the current HEAD of git, in a state in between
relases.

So, 5.6.0 is a release with a tag, and 5.6.2 will be a release with a
tag, but 5.6.1 is just the intermediate state of the tree between
those two releases.

https://en.wikipedia.org/wiki/Software_versioning#Odd-numbered_versions_for_development_releases
https://stackoverflow.com/questions/1266411/which-open-source-projects-use-odd-unstable-even-stable-style-of-versioning